Move models and views in seperate files in projects django app
