+++ rdmo/projects/views/__init__.py
@@ -1,0 +1,15 @@
+from .issue import IssueSendView, IssueUpdateView
+from .membership import (MembershipCreateView, MembershipDeleteView,
+                         MembershipUpdateView)
+from .project import (ProjectDeleteView, ProjectDetailView, ProjectErrorView,
+                      ProjectExportView, ProjectQuestionsView, ProjectsView,
+                      SiteProjectsView)
+from .project_answers import ProjectAnswersExportView, ProjectAnswersView
+from .project_create import (ProjectCreateImportView, ProjectCreateUploadView,
+                             ProjectCreateView)
+from .project_update import (ProjectUpdateImportView, ProjectUpdateTasksView,
+                             ProjectUpdateUploadView, ProjectUpdateView,
+                             ProjectUpdateViewsView)
+from .project_view import ProjectViewExportView, ProjectViewView
+from .snapshot import (SnapshotCreateView, SnapshotRollbackView,
+                       SnapshotUpdateView)
