+++ rdmo/projects/views/project_view.py
@@ -1,0 +1,80 @@
+import logging
+
+from django.conf import settings
+from django.http import Http404
+from django.template import TemplateSyntaxError
+from django.views.generic import DetailView
+
+from rdmo.core.utils import render_to_format
+from rdmo.core.views import ObjectPermissionMixin
+from rdmo.views.models import View
+
+from ..models import Project, Snapshot
+
+logger = logging.getLogger(__name__)
+
+
+class ProjectViewView(ObjectPermissionMixin, DetailView):
+    model = Project
+    queryset = Project.objects.all()
+    permission_required = 'projects.view_project_object'
+    template_name = 'projects/project_view.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(ProjectViewView, self).get_context_data(**kwargs)
+
+        try:
+            context['current_snapshot'] = context['project'].snapshots.get(pk=self.kwargs.get('snapshot_id'))
+        except Snapshot.DoesNotExist:
+            context['current_snapshot'] = None
+
+        try:
+            context['view'] = context['project'].views.get(pk=self.kwargs.get('view_id'))
+        except View.DoesNotExist:
+            raise Http404
+
+        try:
+            context['rendered_view'] = context['view'].render(context['project'], context['current_snapshot'])
+        except TemplateSyntaxError:
+            context['rendered_view'] = None
+
+        context.update({
+            'snapshots': list(context['project'].snapshots.values('id', 'title')),
+            'export_formats': settings.EXPORT_FORMATS
+        })
+
+        return context
+
+
+class ProjectViewExportView(ObjectPermissionMixin, DetailView):
+    model = Project
+    queryset = Project.objects.all()
+    permission_required = 'projects.view_project_object'
+
+    def get_context_data(self, **kwargs):
+        context = super(ProjectViewExportView, self).get_context_data(**kwargs)
+
+        try:
+            context['current_snapshot'] = context['project'].snapshots.get(pk=self.kwargs.get('snapshot_id'))
+        except Snapshot.DoesNotExist:
+            context['current_snapshot'] = None
+
+        try:
+            context['view'] = context['project'].views.get(pk=self.kwargs.get('view_id'))
+        except View.DoesNotExist:
+            raise Http404
+
+        try:
+            context['rendered_view'] = context['view'].render(context['project'], context['current_snapshot'])
+        except TemplateSyntaxError:
+            context['rendered_view'] = None
+
+        context.update({
+            'format': self.kwargs.get('format'),
+            'title': context['project'].title
+        })
+
+        return context
+
+    def render_to_response(self, context, **response_kwargs):
+        return render_to_format(self.request, context['format'], context['title'], 'projects/project_view_export.html', context)
